# swagger.json
{
   "swagger": "2.0",
   "info": {
      "title": "AutoMart API Documentation",
      "description": "This is the documentation for Auto Mart API, an online marketplace where users can buy and sell cars of diverse makes.",
      "version": "1.1.0"
   },
   "host": "auto-mart-api.herokuapp.com",
   "basePath": "/api/v1",
   "schemes": [
      "https"
   ],
...
   "definitions": {
...
      "Car": {
         "type": "object",
         "properties": {
            "_id": {
               "type": "string",
               "description": "The car ID."
            },
            "owner_id": {
               "type": "string",
               "description": "The seller ID."
            },
            "name": {
               "type": "string",
               "description": "The car name."
            },
            "description": {
               "type": "string",
               "description": "The car description."
            },
            "state": {
               "type": "string",
               "description": "The state of the car (new or used)."
            },
            "status": {
               "type": "string",
               "description": "The status of the car (available or unavailable)."
            },
            "price": {
               "type": "number",
               "description": "The price of the car."
            },
            "manufacturer": {
               "type": "string",
               "description": "The makers of the car."
            },
            "model": {
               "type": "string",
               "description": "The car model."
            },
            "body_type": {
               "type": "string",
               "description": "The car body type."
            },
            "image": {
               "properties": {
                  "items": {
                     "type": "array",
                     "items": {}
                  }
               },
               "description": "The car image details containing the image url and id."
            },
            "date": {
               "type": "string",
               "format": "date-time",
               "description": "The listing date of the car."
            }
         }
      },
      "Order": {
         "type": "object",
         "properties": {
            "_id": {
               "type": "string",
               "description": "The order ID."
            },
            "buyer_id": {
               "type": "string",
               "description": "The ID of the buyer."
            },
            "buyer_email": {
               "type": "string",
               "description": "The email of the buyer."
            },
            "car_id": {
               "type": "string",
               "description": "The ID of the car."
            },
            "anount": {
               "type": "number",
               "description": "The order amount."
            },
            "status": {
               "type": "string",
               "description": "The order status."
            },
            "date": {
               "type": "string",
               "format": "date-time",
               "description": "The order creation date."
            }
         }
      }
   },
   "securityDefinitions": {
      "BearerAuth": {
         "type": "apiKey",
         "in": "header",
         "name": "Authorization"
      }
   },
...
   "paths": {
...
      "/car/{car_id}": {
         "get": {
            "description": "Returns a specific car matching the id.",
            "tags": [
               "car"
            ],
            "parameters": [
               {
                  "name": "car_id",
                  "in": "path",
                  "description": "5da1b7fc6c649900174252ea",
                  "type": "string",
                  "required": true
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "schema": {
                     "$ref": "#/definitions/Car"
                  }
               }
            }
         }
      },
...
      "/car/{car_id}/": {
         "delete": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Deletes a car.",
            "tags": [
               "car"
            ],
            "parameters": [
               {
                  "name": "car_id",
                  "in": "path",
                  "description": "5da25a1d0fa5bf00171f96d2",
                  "type": "string",
                  "required": true
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK"
               }
            }
         }
      },
      "/order/car/{car_id}": {
         "get": {
            "description": "Returns all pending orders for a specific car.",
            "tags": [
               "order"
            ],
            "parameters": [
               {
                  "name": "car_id",
                  "in": "path",
                  "description": "5da1b78a6c649900174252e8",
                  "type": "string",
                  "required": true
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "schema": {
                     "$ref": "#/definitions/Order"
                  }
               }
            }
         }
      },
      "/order/user": {
         "get": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Returns all orders made by currently logged in user.",
            "tags": [
               "order"
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "schema": {
                     "$ref": "#/definitions/Order"
                  }
               }
            }
         }
      },
      "/order": {
         "post": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Creates a new order for a car.",
            "tags": [
               "order"
            ],
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "parameters": [
               {
                  "in": "query",
                  "name": "car_id",
                  "description": "Car Id e.g, '5da1b78a6c649900174252e8'",
                  "type": "string",
                  "required": true
               },
               {
                  "in": "body",
                  "name": "New Order",
                  "description": "Enter order amount.",
                  "schema": {
                     "type": "object",
                     "required": [
                        "amount"
                     ],
                     "properties": {
                        "amount": {
                           "type": "number"
                        }
                     }
                  }
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "schema": {
                     "$ref": "#/definitions/Order"
                  }
               }
            }
         }
      },
      "/order/{order_id}/status": {
         "patch": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Marks an order/offer as accepted.",
            "tags": [
               "order"
            ],
            "parameters": [
               {
                  "name": "order_id",
                  "in": "path",
                  "description": "Order Id to accept e.g, '5da25a1d0fa5bf00171f96d2'",
                  "type": "string",
                  "required": true
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK"
               }
            }
         }
      },
      "/order/{order_id}/amount": {
         "patch": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Updates the order amountr.",
            "tags": [
               "order"
            ],
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "parameters": [
               {
                  "name": "order_id",
                  "in": "path",
                  "description": "5da25a1d0fa5bf00171f96d2",
                  "type": "string",
                  "required": true
               },
               {
                  "in": "body",
                  "name": "Update order amount",
                  "description": "Enter the new order amount.",
                  "schema": {
                     "type": "object",
                     "required": [
                        "new_amount"
                     ],
                     "properties": {
                        "new_amount": {
                           "type": "number"
                        }
                     }
                  }
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK"
               }
            }
         }
      },
      "/order/{order_id}/": {
         "delete": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Deletes a pending order.",
            "tags": [
               "order"
            ],
            "parameters": [
               {
                  "name": "order_id",
                  "in": "path",
                  "description": "5da25a1d0fa5bf00171f96d2",
                  "type": "string",
                  "required": true
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK"
               }
            }
         }
      }
   }
}